Refactor implementation of feature flag for Comment Threading
