+++ Library/OptimizelyFeature+Helpers.swift
@@ -1,0 +1,6 @@
+import KsApi
+
+public func featureCommentThreadingIsEnabled() -> Bool {
+  AppEnvironment.current.optimizelyClient?
+    .isFeatureEnabled(featureKey: OptimizelyFeature.Key.commentThreading.rawValue) ?? false
+}
